ref: split repository models
