+++ src/nexuscli/repository/__init__.py
@@ -1,2 +1,3 @@
-from .model import Repository, RepositoryCollection, REMOTE_PATH_SEPARATOR
+from .model import Repository, REMOTE_PATH_SEPARATOR
+from .collection import RepositoryCollection
 from . import validations
